requirement.txt
langchain
python-dotenv
langchain-openai
langchain_community
faiss-cpu
streamlit
langchain-groq
langchain-chroma
